fix mac unitest; test=develop

+++ paddle/fluid/operators/conv_cudnn_helper.h
@@ -92,32 +92,46 @@ std::ostream& operator<<(std::ostream& out, const std::vector<T>& v) {
 
 using framework::AlgorithmsCache;
 
-class ConvSearchCache
-{
-    public:
-        static ConvSearchCache& Instance(){
-            static ConvSearchCache instance;
-            return instance;
-        }
+class ConvSearchCache {
+ public:
+  static ConvSearchCache& Instance() {
+    static ConvSearchCache instance;
+    return instance;
+  }
 
-        framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t>* GetForward() { return &forward_cache_; }
-        framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t>* GetBackwardData() { return &backward_data_cache_; }
-        framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t>* GetBackwardFilter() { return &backward_filter_cache_; }
-        framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t>* GetConvFusion() { return &fusion_forward_cache_; }
-    private:
-        ConvSearchCache() {     }
-        ~ConvSearchCache() {}
-        ConvSearchCache( const ConvSearchCache& ){}
-        ConvSearchCache& operator=( const ConvSearchCache& ) {}
-        framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t> forward_cache_;
-        framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t> backward_data_cache_;
-        framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t> backward_filter_cache_;
-        framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t> fusion_forward_cache_;
+  framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t>* GetForward() {
+    return &forward_cache_;
+  }
+  framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t>* GetBackwardData() {
+    return &backward_data_cache_;
+  }
+  framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t>*
+  GetBackwardFilter() {
+    return &backward_filter_cache_;
+  }
+  framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t>* GetConvFusion() {
+    return &fusion_forward_cache_;
+  }
+
+ private:
+  ConvSearchCache() {}
+  ~ConvSearchCache() {}
+  ConvSearchCache(const ConvSearchCache&) {}
+  ConvSearchCache& operator=(const ConvSearchCache&) {}
+  framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t> forward_cache_;
+  framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t>
+      backward_data_cache_;
+  framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t>
+      backward_filter_cache_;
+  framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t> fusion_forward_cache_;
 };
 
-//static framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t> global_forward_cache;
-//static framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t> global_bwddata_cache;
-//static framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t> global_bwdfilter_cache;
+// static framework::AlgorithmsCache<cudnnConvolutionFwdAlgo_t>
+// global_forward_cache;
+// static framework::AlgorithmsCache<cudnnConvolutionBwdDataAlgo_t>
+// global_bwddata_cache;
+// static framework::AlgorithmsCache<cudnnConvolutionBwdFilterAlgo_t>
+// global_bwdfilter_cache;
 
 struct ConvArgs {
   cudnnHandle_t handle;
@@ -152,7 +166,7 @@ struct SearchAlgorithm<cudnnConvolutionFwdAlgoPerf_t> {
   template <typename T>
   static algo_t Find(const ConvArgs& args, bool exhaustive_search,
                      bool deterministic, int algo_cache_id,
-                     const framework::ExecutionContext& ctx ) {
+                     const framework::ExecutionContext& ctx) {
     auto dtype = platform::CudnnDataType<T>::type;
     bool has_got_workspace_size = true;
     bool exhaustive = (exhaustive_search) & (dtype != CUDNN_DATA_HALF);
@@ -207,16 +221,18 @@ struct SearchAlgorithm<cudnnConvolutionFwdAlgoPerf_t> {
 #endif
       VLOG(3) << "choose algo " << algo;
     } else {
-      //AlgorithmsCache<algo_t>& algo_cache =
+      // AlgorithmsCache<algo_t>& algo_cache =
       //    ctx.GetKernelConfig<AlgorithmsCache<algo_t>>(algo_cache_id);
       auto& dev_ctx =
           ctx.template device_context<platform::CUDADeviceContext>();
       auto workspace_handle = dev_ctx.cudnn_workspace_handle();
-    
+
       auto& temp = ctx.cuda_device_context();
-      //AlgorithmsCache<algo_t>& algo_cache = temp.GetKernelList<AlgorithmsCache<algo_t>>( ctx.Type(), 
+      // AlgorithmsCache<algo_t>& algo_cache =
+      // temp.GetKernelList<AlgorithmsCache<algo_t>>( ctx.Type(),
       //        framework::DataLayout::kNCHW, algo_cache_id);
-      AlgorithmsCache<algo_t>& algo_cache = *(ConvSearchCache::Instance().GetForward());
+      AlgorithmsCache<algo_t>& algo_cache =
+          *(ConvSearchCache::Instance().GetForward());
 
       auto x_dims = framework::vectorize(args.x->dims());
       auto w_dims = framework::vectorize(args.w->dims());
@@ -227,7 +243,8 @@ struct SearchAlgorithm<cudnnConvolutionFwdAlgoPerf_t> {
                << args.p << ", args.d" << args.d;
 
       algo = algo_cache.GetAlgorithm(
-          x_dims, w_dims, args.s, args.p, args.d, 0, args.cudnn_dtype, [&]() {
+          x_dims, w_dims, args.s, args.p, args.d, 0,
+          static_cast<int64_t>(args.cudnn_dtype), [&]() {
             int returned_algo_count;
             std::array<perf_t, kNUM_CUDNN_FWD_ALGS> perf_stat;
 
@@ -273,7 +290,7 @@ struct SearchAlgorithm<cudnnConvolutionBwdDataAlgoPerf_t> {
   template <typename T>
   static algo_t Find(const ConvArgs& args, bool exhaustive_search,
                      bool deterministic, int algo_cache_id,
-                     const framework::ExecutionContext& ctx ) {
+                     const framework::ExecutionContext& ctx) {
     auto dtype = platform::CudnnDataType<T>::type;
     bool exhaustive = (exhaustive_search) & (dtype != CUDNN_DATA_HALF);
     size_t workspace_size_limit = FLAGS_conv_workspace_size_limit * 1024 * 1024;
@@ -345,13 +362,14 @@ struct SearchAlgorithm<cudnnConvolutionBwdDataAlgoPerf_t> {
     } else if (deterministic) {
       return CUDNN_CONVOLUTION_BWD_DATA_ALGO_1;
     } else {
-      //AlgorithmsCache<algo_t>& algo_cache =
+      // AlgorithmsCache<algo_t>& algo_cache =
       //    ctx.GetKernelConfig<AlgorithmsCache<algo_t>>(algo_cache_id);
       auto& dev_ctx =
           ctx.template device_context<platform::CUDADeviceContext>();
       auto workspace_handle = dev_ctx.cudnn_workspace_handle();
-        
-      AlgorithmsCache<algo_t>& algo_cache = *(ConvSearchCache::Instance().GetBackwardData());
+
+      AlgorithmsCache<algo_t>& algo_cache =
+          *(ConvSearchCache::Instance().GetBackwardData());
       /*
       AlgorithmsCache<algo_t>& algo_cache =
           dev_ctx.GetKernelList<AlgorithmsCache<algo_t>>( ctx.Type(),
@@ -368,7 +386,8 @@ struct SearchAlgorithm<cudnnConvolutionBwdDataAlgoPerf_t> {
                << args.p << ", args.d" << args.d;
 
       algo = algo_cache.GetAlgorithm(
-          x_dims, w_dims, args.s, args.p, args.d, 0, args.cudnn_dtype, [&]() {
+          x_dims, w_dims, args.s, args.p, args.d, 0,
+          static_cast<int64_t>(args.cudnn_dtype), [&]() {
             int returned_algo_count;
             std::array<perf_t, kNUM_CUDNN_FWD_ALGS> perf_stat;
 
@@ -418,7 +437,7 @@ struct SearchAlgorithm<cudnnConvolutionBwdFilterAlgoPerf_t> {
   template <typename T>
   static algo_t Find(const ConvArgs& args, bool exhaustive_search,
                      bool deterministic, int algo_cache_id,
-                     const framework::ExecutionContext& ctx ) {
+                     const framework::ExecutionContext& ctx) {
     auto dtype = platform::CudnnDataType<T>::type;
     bool exhaustive = (exhaustive_search) & (dtype != CUDNN_DATA_HALF);
     size_t workspace_size_limit = FLAGS_conv_workspace_size_limit * 1024 * 1024;
@@ -479,7 +498,7 @@ struct SearchAlgorithm<cudnnConvolutionBwdFilterAlgoPerf_t> {
     } else if (deterministic) {
       return CUDNN_CONVOLUTION_BWD_FILTER_ALGO_1;
     } else {
-      //AlgorithmsCache<algo_t>& algo_cache =
+      // AlgorithmsCache<algo_t>& algo_cache =
       //    ctx.GetKernelConfig<AlgorithmsCache<algo_t>>(algo_cache_id);
       auto& dev_ctx =
           ctx.template device_context<platform::CUDADeviceContext>();
@@ -490,7 +509,8 @@ struct SearchAlgorithm<cudnnConvolutionBwdFilterAlgoPerf_t> {
                   framework::DataLayout::kNCHW,
                               algo_cache_id);
      */
-      AlgorithmsCache<algo_t>& algo_cache = *(ConvSearchCache::Instance().GetBackwardFilter());
+      AlgorithmsCache<algo_t>& algo_cache =
+          *(ConvSearchCache::Instance().GetBackwardFilter());
 
       auto x_dims = framework::vectorize(args.x->dims());
       auto w_dims = framework::vectorize(args.w->dims());
@@ -501,7 +521,8 @@ struct SearchAlgorithm<cudnnConvolutionBwdFilterAlgoPerf_t> {
                << args.p << ", args.d" << args.d;
 
       algo = algo_cache.GetAlgorithm(
-          x_dims, w_dims, args.s, args.p, args.d, 0, args.cudnn_dtype, [&]() {
+          x_dims, w_dims, args.s, args.p, args.d, 0,
+          static_cast<int64_t>(args.cudnn_dtype), [&]() {
             int returned_algo_count;
             std::array<perf_t, kNUM_CUDNN_FWD_ALGS> perf_stat;
             auto cudnn_find_func = [&](void* cudnn_workspace_ptr) {
